// src/app/model/intervalsInfoRequest.ts
/**
 * Oil and Gas Reporting System (OGRS-API)
 * This is an api  to allows users to obtain infomation about the oil and gas wells database and generating a punch of related reposts such as wells count, insurance as well as production allocation funcationality
 *
 * OpenAPI spec version: 3.0
 * Contact: mohamed.wafa770@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IntervalsInfoRequest { 
    wellId?: number;
    drivingMechanism?: IntervalsInfoRequest.DrivingMechanismEnum;
    zones?: string;
    interval?: string;
    from?: number;
    to?: number;
    length?: number;
    pr?: number;
    tr?: number;
    status?: IntervalsInfoRequest.StatusEnum;
    startDate?: Date;
    endDate?: Date;
    pi?: number;
    mobility?: number;
    fluidType?: string;
    api?: number;
    notes?: string;
}
export namespace IntervalsInfoRequest {
    export type DrivingMechanismEnum = 'WaterDrive' | 'GasCap' | 'DepleationDrive' | 'GravityDrainage' | 'CombinationDrive';
    export const DrivingMechanismEnum = {
        WaterDrive: 'WaterDrive' as DrivingMechanismEnum,
        GasCap: 'GasCap' as DrivingMechanismEnum,
        DepleationDrive: 'DepleationDrive' as DrivingMechanismEnum,
        GravityDrainage: 'GravityDrainage' as DrivingMechanismEnum,
        CombinationDrive: 'CombinationDrive' as DrivingMechanismEnum
    };
    export type StatusEnum = 'active' | 'shutin';
    export const StatusEnum = {
        Active: 'active' as StatusEnum,
        Shutin: 'shutin' as StatusEnum
    };
}